Bugfix resource loading
If you wonder why I made the hml file a one liner... It is because when coping the whole table from modor for example to google tables. GTables does not like Tabs or newlines it interprets that wrongly and as html does not care I thought I would quickly fix that that way.
Fixes #63